# doc/swagger-admin.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VEGE",
    "description": "E commerce",
    "version": "1.0.0"
  },
  "basePath": "/",
...
  "paths": {
...
    "/admin/list-category?limit=5&offset=1": {
      "get": {
        "summary": "Category List",
        "operationId": "categoryList",
        "tags": ["Category"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request. Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}